requirements.txt
pyspark==3.5.0
pandas==2.1.0
matplotlib==3.7.2
kaggle==1.5.13
python-dotenv==1.0.0
jupyter==1.0.0
numpy==1.24.3 